The method lock_area should call Srv_LockArea (#83)
* Test that we can lock an area

Add unit test that fails if we fail to lock the area. The test relies on
threads and waiting. The guess is that one second is enough time to
prove that the thread calling lock the second time is blocked. This is
could be very fragile.

* The method lock_area should call Srv_LockArea

The method lock_area called Srv_UnlockArea, which causes us never to
lock an area. This change corrects that.