add support to read/write space-padded fixed-length strings (#378)
these strings are not part of the official s7comm specification, but
they are commonly found in real-world systems. they simply consist of a
byte array that contains characters forming a string, padded on the
right with spaces (ascii 32) when the string length is shorter than the
array.

this commit adds support for reading and writing, as well as usage in
tables (snap7.util.DB), and unit tests.