add support to the util.DB class to read and write an entire table
util.DB is a wrapper class to manipulate data in tabular format. before
this commit, it was not possible to read or write the entire table in
one go. this commit introduces support for such operations.

the rationale of these operations is to have the ability to operate on
the entire table in the same way it is possible to operate on the single
rows, without requiring external iteration.