DB usability improvements (#385)
* add support to the util.DB class to read and write an entire table

util.DB is a wrapper class to manipulate data in tabular format. before
this commit, it was not possible to read or write the entire table in
one go. this commit introduces support for such operations.

the rationale of these operations is to have the ability to operate on
the entire table in the same way it is possible to operate on the single
rows, without requiring external iteration.

* implement more dict-like semantics for util.DB

implement keys(), items(), __contains__() methods for the DB class with
dict-like semantics. implement an export() method, flattening the entire
structure (DB and DB_Row's), with semantics akin to DB_Row.export().

document in more detail the existing dict-like semantics of the DB
class, detailing in particular the instances where the existing
behaviour diverges from Python dicts.

introduce an unit test for the new DB.export() method.

* use correct type hints for util.DB constructor

the existing type hints make some attributes Optional, while a None
value for any of those attribute would be in fact an error.

there might be more esuch cases in the DB class: this commit is not
meant to be exhaustive.